% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ts-package}
\alias{blockr.ts}
\alias{blockr.ts-package}
\title{blockr.ts: Time Series Data Blocks for blockr}
\description{
Time series analysis and transformation blocks for blockr. Provides specialized blocks for working with time series data using tsbox for format conversion and analysis.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com}

}
\keyword{internal}
