% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-airpassenger-block.R
\name{new_ts_airpassenger_block}
\alias{new_ts_airpassenger_block}
\title{AirPassengers Time Series Data Block}
\usage{
new_ts_airpassenger_block(...)
}
\arguments{
\item{...}{Additional arguments passed to new_ts_data_block}
}
\description{
A simple time series data block that displays AirPassengers data as an interactive dygraph.
This block demonstrates time series visualization using tsbox and dygraphs.
}
