% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-bind-block.R
\name{new_ts_bind_block}
\alias{new_ts_bind_block}
\title{Time Series Bind Block}
\usage{
new_ts_bind_block(...)
}
\arguments{
\item{...}{Named time series data to combine. Each argument should be a
time series that can be converted to tsbox format.}
}
\value{
A ts_bind_block object
}
\description{
Combine multiple time series into a single multivariate time series
}
\details{
This block uses tsbox::ts_c() to combine multiple univariate or multivariate
time series into a single multivariate series. Each input series becomes a
separate series in the output, identified by the 'id' column.

You can pass multiple time series as named arguments:
\code{new_ts_bind_block(male = mdeaths, female = fdeaths)}
}
