% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-blocks.R
\name{new_ts_data_block}
\alias{new_ts_data_block}
\title{Time Series Data Block Constructor}
\usage{
new_ts_data_block(server, ui, class, ctor = sys.parent(), ...)
}
\arguments{
\item{server}{Server function for the block}

\item{ui}{UI function for the block}

\item{class}{Additional class names for the block}

\item{ctor}{The constructor function environment}

\item{...}{Additional arguments passed to new_data_block}
}
\description{
Base constructor for time series data blocks that automatically display results as dygraphs.
This adds both ts_data_block and ts_block classes to ensure dygraph rendering.
}
