% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-dataset-block.R
\name{new_ts_dataset_block}
\alias{new_ts_dataset_block}
\title{Time Series Dataset Selector Block}
\usage{
new_ts_dataset_block(dataset = "AirPassengers", ...)
}
\arguments{
\item{dataset}{Name of the dataset to load (default: "AirPassengers")}

\item{...}{Additional arguments passed to new_ts_data_block}
}
\description{
A data block that provides access to all built-in R time series datasets.
Supports both univariate and multivariate time series, automatically converting
them to tidy data frames using tsbox::ts_tbl().
}
