% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-forecast-block.R
\name{new_ts_forecast_block}
\alias{new_ts_forecast_block}
\title{Time Series Forecasting Block}
\usage{
new_ts_forecast_block(horizon = 12, ...)
}
\arguments{
\item{horizon}{Integer. Number of periods to forecast (default: 12).}

\item{...}{Additional arguments passed to new_ts_transform_block()}
}
\value{
A ts_forecast_block object
}
\description{
Generate forecasts with confidence intervals
}
