% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-lag-block.R
\name{new_ts_lag_block}
\alias{new_ts_lag_block}
\title{Time Series Lag/Lead Block}
\usage{
new_ts_lag_block(by = 1L, ...)
}
\arguments{
\item{by}{Integer specifying the number of periods to shift:
\itemize{
\item Positive values create lags (shift forward in time)
\item Negative values create leads (shift backward in time)
\item Default is 1 (one period lag)
}}

\item{...}{Additional arguments passed to new_ts_transform_block}
}
\description{
A transform block that shifts time series forward (lag) or backward (lead).
}
\details{
This block uses tsbox::ts_lag() to shift time series data by a specified number
of periods. Lagging is useful for comparing current values with past values,
while leading can be used for forecasting comparisons.
}
