% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-pc-block.R
\name{new_ts_pc_block}
\alias{new_ts_pc_block}
\title{Time Series Percentage Change Transform Block}
\usage{
new_ts_pc_block(...)
}
\arguments{
\item{...}{Additional arguments passed to new_ts_transform_block}
}
\description{
A simple transform block that computes percentage changes for time series data.
Uses tsbox::ts_pc() to calculate the percentage change between consecutive observations.
}
\details{
This block:
\itemize{
\item Takes time series data as input (from a ts data block)
\item Applies tsbox::ts_pc() to compute percentage changes
\item Returns data where each value is the percentage change from the previous value
\item The first value will be NA (no previous value to compare)
\item Works with both univariate and multivariate time series
}
}
