% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-span-block.R
\name{new_ts_span_block}
\alias{new_ts_span_block}
\title{Time Series Span (Time Range) Block}
\usage{
new_ts_span_block(start = NULL, end = NULL, ...)
}
\arguments{
\item{start}{Character string or Date specifying the start date.
Use NULL to start from the beginning of the series.}

\item{end}{Character string or Date specifying the end date.
Use NULL to include up to the end of the series.}

\item{...}{Additional arguments passed to new_ts_transform_block}
}
\description{
A transform block that filters time series to a specific time range.
}
\details{
This block uses tsbox::ts_span() to filter time series data to a specific
time window. The dates can be specified in various formats that tsbox
understands (e.g., "2020-01-01", "2020-01", "2020").
}
