% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-select-block.R
\name{new_ts_select_block}
\alias{new_ts_select_block}
\title{Time Series Selection Block}
\usage{
new_ts_select_block(series = NULL, ...)
}
\arguments{
\item{series}{Character vector of series names to select (default: NULL selects all)}

\item{...}{Additional arguments passed to new_ts_transform_block}
}
\description{
A transform block that selects specific series from multivariate time series data.
}
\details{
This block uses tsbox::ts_pick() to select specific series from multivariate
time series data by explicitly selecting series by name.
If series is NULL or empty, all available series are selected.
}
